--- converted_tflite/labels.txt
@@ -1,2 +1,0 @@
-0 Class 1
-1 Class 2
