+++ main.py
@@ -1,114 +1,196 @@
-#!/usr/bin/env python3
-
-from camera import *
-from ai_magic import Ai
-#from lcd_driver import Display
-import time
-#import RPi.GPIO as GPIO
-#import signal
-import sys
-import cv2
+import multiprocessing
 import numpy as np
+import cv2
+import keras as tf
+import pyttsx3
+import math
+import os
+
+DIR_PATH = os.path.dirname(os.path.realpath(__file__))
 
 
-ai = Ai('converted_tflite')
-"""lcd = Display()
-BUTTON_GPIO = 26
-run = False"""
-
-"""def signal_handler(sig, frame):
-    GPIO.cleanup()
-    sys.exit(0)"""
-
-"""
-def button_pressed_callback(channel):
-    print('Click')
-    lcd.clear()
-    image = camera.get_frame()
-    ai.classify_image(image)
-    print(ai.get_prediction())
-    lcd.write(str(ai.get_prediction()[0]))
-    lcd.new_line()
-    lcd.write(str(ai.get_prediction()[1]))
-"""
-
-def visualize(image):
-    """
-    for i in range(len(prediction)):
-        image = np.array(image)
-
-        color = (0, 0, 255)
-        boxes = ai.bbox()
-        imW, imH = int(width), int(height)
-        #boxes = boxes
-
-        ymin = int(max(1, (boxes[i][0] * imH)))
-        xmin = int(max(1, (boxes[i][1] * imW)))
-        ymax = int(min(imH, (boxes[i][2] * imH)))
-        xmax = int(min(imW, (boxes[i][3] * imW)))
-
-        cv2.rectangle(image, (xmin, ymin), (xmax, ymax), color, 2)
-    """
-    prediction = ai.classify_image(image)
-    image = np.array(image)
-    image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
-    
-    position = np.squeeze(ai.positions)
-    for i in range(len(ai.positions)):
-        color = (0, 0, 255)
-        imW, imH = int(width), int(height)
-        
-        print(ai.get_prediction())
-        print(position)
-        """
-        ymin = int(1 * (position[0] * imH))
-        xmin = int(1 * (position[1] * imW))
-        ymax = int(ymin * (position[0] * imH))
-        xmax = int(xmin * (position[1] * imW))
-        """
-        ymin = int(position[1] * imH)
-        xmin = int(position[0] * imW)
-        ymax = int(position[1] * imH + ymin)
-        xmax = int(position[0]* imW + xmin)
-        
-        print(xmin, ymin, xmax, ymax)
-
-        #cv2.rectangle(image, (xmax, ymax), (xmin, ymin), color, 2)
-        cv2.rectangle(image, (xmin, ymin), (xmax, ymax), color, 2)
-        cv2.imshow("image",image) 
-    #cv2.imshow("image",image)
-    
+def speak(speakQ, ):
+    engine = pyttsx3.init()
+    volume = engine.getProperty('volume')
+    engine.setProperty('volume', volume)
+    last_msg = ""
+    while True:
+        msg = speakQ.get()
+        while not speakQ.empty():
+            msg = speakQ.get()
+        if msg != last_msg and msg != "Background":
+            last_msg = msg
+            engine.say(msg)
+            engine.runAndWait()
+        if msg == "Background":
+            last_msg = ""
 
 
 def main():
-    while True:
-        image = get_image_stream()
-        #prediction = ai.classify_image(image)
-        #print(ai.get_prediction()) #DEV STUFF comment all line in final version
-        #print(ai.positions) #DEV STUFF comment all line in final version
-        visualize(image)
+    labels_path = f"{DIR_PATH}/model/labels.txt"
+    labelsfile = open(labels_path, 'r')
 
-        if cv2.waitKey(1) & 0xFF == ord("q"):
-            break
+    classes = []
+    line = labelsfile.readline()
+    while line:
+        classes.append(line.split(' ', 1)[1].rstrip())
+        line = labelsfile.readline()
+    labelsfile.close()
+
+    model_path = f"{DIR_PATH}/model/keras_model.h5"
+    model = tf.models.load_model(model_path, compile=False)
+
+    cap = cv2.VideoCapture(0)
+
+    frameWidth = 1024
+    frameHeight = 768
+
+    cap.set(cv2.CAP_PROP_FRAME_WIDTH, frameWidth)
+    cap.set(cv2.CAP_PROP_FRAME_HEIGHT, frameHeight)
+    # enable auto gain
+    cap.set(cv2.CAP_PROP_GAIN, 0)
+    
+    # creating a queue to share data to speech process
+    speakQ = multiprocessing.Queue()
+
+    # creating speech process to not hang processor
+    p1 = multiprocessing.Process(target=speak, args=(speakQ, ), daemon="True")
+
+    # starting process 1 - speech
+    p1.start()
+    
+
+    while True:
+
+        # disable scientific notation for clarity
+        np.set_printoptions(suppress=True)
+
+        # Create the array of the right shape to feed into the keras model.
+        # We are inputting 1x 224x224 pixel RGB image.
+        data = np.ndarray(shape=(1, 224, 224, 3), dtype=np.float32)
 
         
+        check, frame = cap.read()
+        frame = cv2.rotate(frame, cv2.ROTATE_90_COUNTERCLOCKWISE)
 
-    print("SUCCESSFULY DONE ;-)")
+        
+        margin = int(((frameWidth-frameHeight)/2))
+        square_frame = frame[0:frameHeight, margin:margin + frameHeight]
+        resized_img = cv2.resize(square_frame, (224, 224))
+        model_img = cv2.cvtColor(resized_img, cv2.COLOR_BGR2RGB)
+
+        image_array = np.asarray(model_img)
+        normalized_image_array = (image_array.astype(np.float32) / 127.0) - 1
+        data[0] = normalized_image_array
+
+        predictions = model.predict(data)
+        conf_threshold = 50
+        confidence = []
+        conf_label = ""
+        threshold_class = ""
+        per_line = 2  
+        bordered_frame = cv2.copyMakeBorder(
+            square_frame,
+            top=0,
+            bottom=30 + 15*math.ceil(len(classes)/per_line),
+            left=0,
+            right=0,
+            borderType=cv2.BORDER_CONSTANT,
+            value=[0, 0, 0]
+        )
+
+        for i in range(0, len(classes)):
+            confidence.append(int(predictions[0][i]*100))
+            if confidence[i] > conf_threshold:
+
+                speakQ.put(classes[i])
+                threshold_class = classes[i]
+                print(threshold_class, " ", confidence[i], "%")
+                
+                conf_label = classes[i] + ": " + str(confidence[i]) + "%; "
+
+                print(conf_label)
+                """
+                cv2.putText(
+                    bordered_frame, 
+                    conf_label, 
+                    (int(0), int(frameHeight-50)),
+                    cv2.FONT_HERSHEY_SIMPLEX,
+                    0.75,
+                    (255, 255, 255)
+                )
+                """
+                
+                cv2.putText(
+                    img=bordered_frame,
+                    text=conf_label,
+                    #org=(int(0), int(-(frameHeight+25+15*math.ceil(i/per_line)))),
+                    org=(120, 240),
+                    fontFace=cv2.FONT_HERSHEY_SIMPLEX,
+                    fontScale=1,
+                    color=(255, 255, 255)
+                )
+                    
+                """
+                conf_label = classes[i] + ": " + str(confidence[i]) + "%; "
+        
+                if (i == (len(classes)-1)):
+                    cv2.putText(
+                        img=bordered_frame,
+                        text=conf_label,
+                        org=(0), int(frameHeight + 25 + 15 * math.ceil((i + 1) / per_line))),
+                        fontFace=cv2.FONT_HERSHEY_SIMPLEX,
+                        fontScale=0.5,
+                        color=(255, 255, 255)
+                    )
+                    conf_label = ""
+                """
+            """
+            cv2.putText(
+                    img=bordered_frame,
+                    text=threshold_class,
+                    org=(int(0), int(frameHeight+20)),
+                    fontFace=cv2.FONT_HERSHEY_SIMPLEX,
+                    fontScale=2,
+                    color=(255, 255, 255)
+            )
+                #print(classes) #DEV STUFF comment all line in final version
+            """         
+
+            """
+            if confidence[i] > conf_threshold:
+                speakQ.put(classes[i])
+                threshold_class = classes[i]
+                print(threshold_class, confidence[i])
+            
+                
+                cv2.putText(
+                    img=bordered_frame,
+                    text=threshold_class,
+                    org=(int(frameWidth / 2), int(frameHeight + 20)),
+                    fontFace=cv2.FONT_HERSHEY_SIMPLEX,
+                    fontScale=0.75,
+                    color=(255, 255, 255)
+                )
+            """ 
+
+            cv2.putText(
+                    bordered_frame, 
+                    conf_label, 
+                    (int(0), int(frameHeight-50)),
+                    cv2.FONT_HERSHEY_SIMPLEX,
+                    0.75,
+                    (255, 255, 255)
+                )
+
+            cv2.namedWindow('Capturing', cv2.WND_PROP_FULLSCREEN)
+            cv2.setWindowProperty('Capturing', cv2.WND_PROP_FULLSCREEN, cv2.WINDOW_FULLSCREEN)
+            cv2.imshow("Capturing", bordered_frame)
+        if cv2.waitKey(1) & 0xFF == ord('q'):
+            break
+
+    p1.terminate()
+
 
 if __name__ == '__main__':
-    """
-    lcd.write('     NVIAS     ')
-    GPIO.setmode(GPIO.BCM)
-    GPIO.setup(BUTTON_GPIO, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-    GPIO.add_event_detect(BUTTON_GPIO, GPIO.FALLING,
-                          callback=button_pressed_callback, bouncetime=500)
-
-    signal.signal(signal.SIGINT, signal_handler)
-    signal.pause()
-    while(True):
-        time.sleep(1)
-    """
     main()
-
-
-
