--- main2.py
@@ -1,333 +1,0 @@
-"""
-import tflite_runtime.interpreter as tflite
-import os
-import argparse
-import cv2
-import numpy as np
-import sys
-import time
-from threading import Thread
-import importlib.util
-
-# Define VideoStream class to handle streaming of video from webcam in separate processing thread
-# Source - Adrian Rosebrock, PyImageSearch: https://www.pyimagesearch.com/2015/12/28/increasing-raspberry-pi-fps-with-python-and-opencv/
-class VideoStream:
-    Camera object that controls video streaming from the Picamera
-    def __init__(self,resolution=(640,480),framerate=30):
-        # Initialize the PiCamera and the camera image stream
-        self.stream = cv2.VideoCapture(0)
-        ret = self.stream.set(cv2.CAP_PROP_FOURCC, cv2.VideoWriter_fourcc(*'MJPG'))
-        ret = self.stream.set(3,resolution[0])
-        ret = self.stream.set(4,resolution[1])
-            
-        # Read first frame from the stream
-        (self.grabbed, self.frame) = self.stream.read()
-
-	# Variable to control when the camera is stopped
-        self.stopped = False
-
-    def start(self):
-	# Start the thread that reads frames from the video stream
-        Thread(target=self.update,args=()).start()
-        return self
-
-    def update(self):
-        # Keep looping indefinitely until the thread is stopped
-        while True:
-            # If the camera is stopped, stop the thread
-            if self.stopped:
-                # Close camera resources
-                self.stream.release()
-                return
-
-            # Otherwise, grab the next frame from the stream
-            (self.grabbed, self.frame) = self.stream.read()
-
-    def read(self):
-	# Return the most recent frame
-        return self.frame
-
-    def stop(self):
-	# Indicate that the camera and thread should be stopped
-        self.stopped = True
-        
-parser = argparse.ArgumentParser()
-parser.add_argument('--model', help='Provide the path to the TFLite file, default is models/model.tflite',
-                    default='converted_tflite/model_unquant.tflite')
-parser.add_argument('--labels', help='Provide the path to the Labels, default is models/labels.txt',
-                    default='converted_tflite/labels.txt')
-parser.add_argument('--threshold', help='Minimum confidence threshold for displaying detected objects',
-                    default=0.5)
-parser.add_argument('--resolution', help='Desired webcam resolution in WxH. If the webcam does not support the resolution entered, errors may occur.',
-                    default='1280x720')
-                    
-args = parser.parse_args()
-
-# PROVIDE PATH TO MODEL DIRECTORY
-PATH_TO_MODEL_DIR = args.model
-
-# PROVIDE PATH TO LABEL MAP
-PATH_TO_LABELS = args.labels
-
-# PROVIDE THE MINIMUM CONFIDENCE THRESHOLD
-MIN_CONF_THRESH = float(args.threshold)
-
-resW, resH = args.resolution.split('x')
-imW, imH = int(resW), int(resH)
-import time
-print('Loading model...', end='')
-start_time = time.time()
-
-# LOAD TFLITE MODEL
-interpreter = tflite.Interpreter(model_path=PATH_TO_MODEL_DIR)
-# LOAD LABELS
-with open(PATH_TO_LABELS, 'r') as f:
-    labels = [line.strip() for line in f.readlines()]
-end_time = time.time()
-elapsed_time = end_time - start_time
-print('Done! Took {} seconds'.format(elapsed_time))
-
-interpreter.allocate_tensors()
-input_details = interpreter.get_input_details()
-output_details = interpreter.get_output_details()
-
-height = input_details[0]['shape'][1]
-width = input_details[0]['shape'][2]
-
-floating_model = (input_details[0]['dtype'] == np.float32)
-
-input_mean = 127.5
-input_std = 127.5
-
-# Initialize frame rate calculation
-frame_rate_calc = 1
-freq = cv2.getTickFrequency()
-print('Running inference for PiCamera')
-# Initialize video stream
-videostream = VideoStream(resolution=(imW,imH),framerate=30).start()
-time.sleep(1)
-
-#for frame1 in camera.capture_continuous(rawCapture, format="bgr",use_video_port=True):
-while True:
-    # Start timer (for calculating frame rate)
-    current_count=0
-    t1 = cv2.getTickCount()
-
-    # Grab frame from video stream
-    frame1 = videostream.read()
-
-    # Acquire frame and resize to expected shape [1xHxWx3]
-    frame = frame1.copy()
-    frame_rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
-    frame_resized = cv2.resize(frame_rgb, (width, height))
-    input_data = np.expand_dims(frame_resized, axis=0)
-
-    # Normalize pixel values if using a floating model (i.e. if model is non-quantized)
-    if floating_model:
-        input_data = (np.float32(input_data) - input_mean) / input_std
-
-    # Perform the actual detection by running the model with the image as input
-    interpreter.set_tensor(input_details[0]['index'],input_data)
-    interpreter.invoke()
-
-    # Retrieve detection results
-    boxes = interpreter.get_tensor(output_details[0]['index'])[0] # Bounding box coordinates of detected objects
-    classes = interpreter.get_tensor(output_details[1]['index'])[0] # Class index of detected objects
-    scores = interpreter.get_tensor(output_details[2]['index'])[0] # Confidence of detected objects
-    #num = interpreter.get_tensor(output_details[3]['index'])[0]  # Total number of detected objects (inaccurate and not needed)
-
-    # Loop over all detections and draw detection box if confidence is above minimum threshold
-    for i in range(len(scores)):
-        if ((scores[i] > MIN_CONF_THRESH) and (scores[i] <= 1.0)):
-
-            # Get bounding box coordinates and draw box
-            # Interpreter can return coordinates that are outside of image dimensions, need to force them to be within image using max() and min()
-            ymin = int(max(1,(boxes[i][0] * imH)))
-            xmin = int(max(1,(boxes[i][1] * imW)))
-            ymax = int(min(imH,(boxes[i][2] * imH)))
-            xmax = int(min(imW,(boxes[i][3] * imW)))
-            
-            cv2.rectangle(frame, (xmin,ymin), (xmax,ymax), (10, 255, 0), 2)
-
-            # Draw label
-            object_name = labels[int(classes[i])] # Look up object name from "labels" array using class index
-            label = '%s: %d%%' % (object_name, int(scores[i]*100)) # Example: 'person: 72%'
-            labelSize, baseLine = cv2.getTextSize(label, cv2.FONT_HERSHEY_SIMPLEX, 0.7, 2) # Get font size
-            label_ymin = max(ymin, labelSize[1] + 10) # Make sure not to draw label too close to top of window
-            cv2.rectangle(frame, (xmin, label_ymin-labelSize[1]-10), (xmin+labelSize[0], label_ymin+baseLine-10), (255, 255, 255), cv2.FILLED) # Draw white box to put label text in
-            cv2.putText(frame, label, (xmin, label_ymin-7), cv2.FONT_HERSHEY_SIMPLEX, 0.7, (0, 0, 0), 2) # Draw label text
-            current_count+=1
-    # Draw framerate in corner of frame
-    cv2.putText(frame,'FPS: {0:.2f}'.format(frame_rate_calc),(15,25),cv2.FONT_HERSHEY_SIMPLEX,1,(0,255,55),2,cv2.LINE_AA)
-    cv2.putText (frame,'Total Detection Count : ' + str(current_count),(15,65),cv2.FONT_HERSHEY_SIMPLEX,1,(0,255,55),2,cv2.LINE_AA)
-    # All the results have been drawn on the frame, so it's time to display it.
-    cv2.imshow('Object Detector', frame)
-
-    # Calculate framerate
-    t2 = cv2.getTickCount()
-    time1 = (t2-t1)/freq
-    frame_rate_calc= 1/time1
-
-    # Press 'q' to quit
-    if cv2.waitKey(1) == ord('q'):
-        break
-
-# Clean up
-cv2.destroyAllWindows()
-videostream.stop()
-print("Done")
-"""
-
-from imutils.video import VideoStream, FPS
-from tflite_runtime.interpreter import Interpreter, load_delegate
-import argparse
-import time
-import cv2
-import re
-from PIL import Image, ImageDraw, ImageFont
-import numpy as np
-
-
-def draw_image(image, results, labels, size):
-    result_size = len(results)
-    for idx, obj in enumerate(results):
-        print(obj)
-        # Prepare image for drawing
-        draw = ImageDraw.Draw(image)
-
-        # Prepare boundary box
-        ymin, xmin, ymax, xmax = obj['bounding_box']
-        xmin = int(xmin * size[0])
-        xmax = int(xmax * size[0])
-        ymin = int(ymin * size[1])
-        ymax = int(ymax * size[1])
-
-        # Draw rectangle to desired thickness
-        for x in range( 0, 4 ):
-            draw.rectangle((ymin, xmin, ymax, xmax), outline=(255, 255, 0))
-
-        # Annotate image with label and confidence score
-        display_str = labels[obj['class_id']] + ": " + str(round(obj['score']*100, 2)) + "%"
-        draw.text((ymin,xmin), display_str, font=ImageFont.truetype("/usr/share/fonts/truetype/piboto/Piboto-Regular.ttf", 20))
-
-        displayImage = np.asarray( image )
-        cv2.imshow('Coral Live Object Detection', displayImage)
-
-
-def load_labels(path):
-    """Loads the labels file. Supports files with or without index numbers."""
-    with open(path, 'r', encoding='utf-8') as f:
-        lines = f.readlines()
-        labels = {}
-        for row_number, content in enumerate(lines):
-            pair = re.split(r'[:\s]+', content.strip(), maxsplit=1)
-            if len(pair) == 2 and pair[0].strip().isdigit():
-                labels[int(pair[0])] = pair[1].strip()
-            else:
-                labels[row_number] = pair[0].strip()
-    return labels
-
-
-def set_input_tensor(interpreter, image):
-    """Sets the input tensor."""
-    tensor_index = interpreter.get_input_details()[0]['index']
-    input_tensor = interpreter.tensor(tensor_index)()[0]
-    input_tensor[:, :] = image
-
-
-def get_output_tensor(interpreter, index):
-    """Returns the output tensor at the given index."""
-    output_details = interpreter.get_output_details()[index]
-    tensor = np.squeeze(interpreter.get_tensor(output_details['index']))
-    return tensor
-
-
-def detect_objects(interpreter, image, threshold):
-    """Returns a list of detection results, each a dictionary of object info."""
-    set_input_tensor(interpreter, image)
-    interpreter.invoke()
-
-    # Get all output details
-    boxes = get_output_tensor(interpreter, 0)
-    classes = get_output_tensor(interpreter, 1)
-    scores = get_output_tensor(interpreter, 2)
-    count = int(get_output_tensor(interpreter, 3))
-
-    results = []
-    for i in range(count):
-        if scores[i] >= threshold:
-            result = {
-                'bounding_box': boxes[i],
-                'class_id': classes[i],
-                'score': scores[i]
-            }
-            results.append(result)
-    return results
-
-
-def make_interpreter(model_file, use_edgetpu):
-    model_file, *device = model_file.split('@')
-    if use_edgetpu:
-        return Interpreter(
-            model_path=model_file,
-            experimental_delegates=[
-                load_delegate('libedgetpu.so.1',
-                {'device': device[0]} if device else {})
-            ]
-        )
-    else:
-        return Interpreter(model_path=model_file)
-
-
-def main():
-    parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    parser.add_argument('-m', '--model', type=str, help='File path of .tflite file.', default='converted_tflite/model.tflite')
-    parser.add_argument('-l', '--labels', type=str, help='File path of labels file.', default='converted_tflite/labels.txt')
-    parser.add_argument('-t', '--threshold', type=float, default=0.4, required=False, help='Score threshold for detected objects.')
-    parser.add_argument('-p', '--picamera', action='store_true', default=False, help='Use PiCamera for image capture')
-    parser.add_argument('-e', '--use_edgetpu', action='store_true', default=False, help='Use EdgeTPU')
-    args = parser.parse_args()
-
-    labels = load_labels(args.labels)
-    interpreter = make_interpreter(args.model, args.use_edgetpu)
-    interpreter.allocate_tensors()
-    _, input_height, input_width, _ = interpreter.get_input_details()[0]['shape']
-
-    # Initialize video stream
-    vs = VideoStream(usePiCamera=args.picamera, resolution=(640, 480))
-    vs.start() # this would close/ release the camera object properly. For vs = VideoStream().start(), vs.stop() would not work
-    time.sleep(1)
-
-    fps = FPS().start()
-
-    while True:
-        try:
-            # Read frame from video
-            screenshot = vs.read()
-            image = Image.fromarray(screenshot)
-            image_pred = image.resize((input_width ,input_height), Image.ANTIALIAS)
-
-            # Perform inference
-            results = detect_objects(interpreter, image_pred, args.threshold)
-            
-            draw_image(image, results, labels, image.size)
-
-            if( cv2.waitKey( 5 ) & 0xFF == ord( 'q' ) ):
-                fps.stop()
-                break
-
-            fps.update()
-        except KeyboardInterrupt:
-            fps.stop()
-            break
-
-    print("Elapsed time: " + str(fps.elapsed()))
-    print("Approx FPS: :" + str(fps.fps()))
-
-    cv2.destroyAllWindows()
-    vs.stop()
-    time.sleep(2)
-
-
-if __name__ == '__main__':
-    main()
