--- readme.MD
@@ -1,37 +1,0 @@
-<h1>Ai game boy</h1>
-<h2>Instalation guide</h2>
-First of all clone the repository by typing to the console 
-
-```
-git clone https://github.com/DavidZah/game_boy_ai.git
-```
-
-You must allow I2C and Camera by
-
-```
-In raspberry -> Raspberry Pi Configuration -> Interfaces -> enable LCD and I2C (reboot will be required)
-```
-
-After that write
-
-```
-cd game_boy_ai
-pip install -r requirements.txt
-```
-
-<h2>How to run it</h2>
-In console write
-
-```
-python3 main.py
-```
-
-<h2>How add new dataset</h2>
-On teacheble machine train model after that click on export model select Tensorflow Lite
-Model conversion type select Floating point after that click on Download my model. 
-The model unzip and copy and replace to folder converted_tflite 
-
-
-
-[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
-Copyright (c) 2021 nvias, z. s.
